Prevent unnecessary redirect on public pages
Opening a new tab checked current user and if not logged in, would
redirect to home. This is necessary in the CMS but in the main client
should only happen for authorised pages like account profile.
Now opening a new tab will otherwise not redirect and will maintain the
current login status.

Fixed expand comment arrows not rendering on some mobile devices
(now uses SVGs instead of HTML entities)